# New cluster shape filter for phase2 + adding PixelPair