--- a/RecoTracker/FinalTrackSelectors/python/earlyGeneralTracks_cfi.py
+++ b/RecoTracker/FinalTrackSelectors/python/earlyGeneralTracks_cfi.py
@@ -107,16 +107,18 @@ trackingPhase2PU140.toReplaceWith(earlyGeneralTracks, _trackListMerger.clone(
                      'lowPtQuadStepTracks',
                      'lowPtTripletStepTracks',
                      'detachedQuadStepTracks',
+                     'pixelPairStepTracks',
                     ],
-    hasSelector = [1,1,1,1,1],
+    hasSelector = [1,1,1,1,1,1],
     indivShareFrac = [1.0,0.16,0.095,0.09,0.09,0.09],
     selectedTrackQuals = cms.VInputTag(cms.InputTag("initialStepSelector","initialStep"),
                                        cms.InputTag("highPtTripletStepSelector","highPtTripletStep"),
                                        cms.InputTag("lowPtQuadStepSelector","lowPtQuadStep"),
                                        cms.InputTag("lowPtTripletStepSelector","lowPtTripletStep"),
                                        cms.InputTag("detachedQuadStep"),
+                                       cms.InputTag("pixelPairStepSelector","pixelPairStep"),
                                        ),
-    setsToMerge = cms.VPSet( cms.PSet( tLists=cms.vint32(0,1,2,3,4), pQual=cms.bool(True) )
+    setsToMerge = cms.VPSet( cms.PSet( tLists=cms.vint32(0,1,2,3,4,5), pQual=cms.bool(True) )
                              ),
     copyExtras = True,
     makeReKeyedSeeds = cms.untracked.bool(False)