--- a/RecoTracker/IterativeTracking/python/PixelPairStep_cff.py
+++ b/RecoTracker/IterativeTracking/python/PixelPairStep_cff.py
@@ -38,15 +38,15 @@ trackingPhase1PU70.toModify(pixelPairStepSeedLayers,
 )
 
 _layerListForPhase2 = ['BPix1+BPix2', 'BPix1+BPix3', 'BPix2+BPix3',
-                       'BPix2+BPix4', 'BPix3+BPix4',
+#                       'BPix2+BPix4', 'BPix3+BPix4',
                        'BPix1+FPix1_pos', 'BPix1+FPix1_neg',
                        'BPix2+FPix1_pos', 'BPix2+FPix1_neg',
-                       'BPix1+FPix2_pos', 'BPix1+FPix2_neg',
-                       'FPix1_pos+FPix2_pos', 'FPix1_neg+FPix2_neg',
-                       'FPix2_pos+FPix3_pos', 'FPix2_neg+FPix3_neg',
-                       'FPix3_pos+FPix4_pos', 'FPix3_neg+FPix4_neg',
-                       'FPix4_pos+FPix5_pos', 'FPix4_neg+FPix5_neg',
-                       'FPix5_pos+FPix6_pos', 'FPix5_neg+FPix6_neg'
+#                       'BPix1+FPix2_pos', 'BPix1+FPix2_neg',
+#                       'FPix1_pos+FPix2_pos', 'FPix1_neg+FPix2_neg',
+#                       'FPix2_pos+FPix3_pos', 'FPix2_neg+FPix3_neg',
+#                       'FPix3_pos+FPix4_pos', 'FPix3_neg+FPix4_neg',
+#                       'FPix4_pos+FPix5_pos', 'FPix4_neg+FPix5_neg',
+#                       'FPix5_pos+FPix6_pos', 'FPix5_neg+FPix6_neg'
 ]
 from Configuration.Eras.Modifier_trackingPhase2PU140_cff import trackingPhase2PU140
 trackingPhase2PU140.toModify(pixelPairStepSeedLayers, 
@@ -117,6 +117,7 @@ trackingPhase1PU70.toReplaceWith(pixelPairStepTrajectoryFilterBase, _pixelPairSt
     constantValueForLostHitsFractionFilter = 0.801,
 ))
 trackingPhase2PU140.toReplaceWith(pixelPairStepTrajectoryFilterBase, _pixelPairStepTrajectoryFilterBase.clone(
+    minimumNumberOfHits = 4,
     maxLostHitsFraction = 1./10.,
     constantValueForLostHitsFractionFilter = 0.701,
 ))
@@ -128,6 +129,14 @@ pixelPairStepTrajectoryFilter = cms.PSet(
         cms.PSet( refToPSet_ = cms.string('pixelPairStepTrajectoryFilterBase')),
     #    cms.PSet( refToPSet_ = cms.string('pixelPairStepTrajectoryFilterShape'))
     ),
+)
+
+
+pixelPairStepTrajectoryFilterInOut = pixelPairStepTrajectoryFilterBase.clone(
+    minimumNumberOfHits = 4,
+    seedExtension = 1,
+    strictSeedExtension = False, # don't allow inactive
+    pixelSeedExtension = False,
 )
 
 
@@ -162,6 +171,15 @@ pixelPairStepTrajectoryBuilder = RecoTracker.CkfPattern.GroupedCkfTrajectoryBuil
     maxPtForLooperReconstruction = cms.double(0.7) 
     )
 trackingLowPU.toModify(pixelPairStepTrajectoryBuilder, maxCand = 2)
+
+trackingPhase2PU140.toModify(pixelPairStepTrajectoryBuilder,
+    inOutTrajectoryFilter = dict(refToPSet_ = "pixelPairStepTrajectoryFilterInOut"),
+    useSameTrajFilter = False,
+    maxCand = 3,
+)
+
+
+
 
 # MAKING OF TRACK CANDIDATES
 import RecoTracker.CkfPattern.CkfTrackCandidates_cfi