… RecoPixelVertexing/PixelLowPtUtilities/data/pixelShape_Phase2Tk.par
… RecoPixelVertexing/PixelLowPtUtilities/data/pixelShape_WP99p5_Phase2Tk.par
… RecoPixelVertexing/PixelLowPtUtilities/data/pixelShape_WP99p9_Phase2Tk.par