This Pr includes three main changes (all connected )
1) add a New cluster shape filter for phase2 
1a) I explicitly added two files for two efficiency WP 99.5 and 99.9 the default (at the moment a copy) is 99.5
2) add (back) a PixelPair iteration to recover any inefficiency  of the triplet step
(so PixelPair cover only the regions not covered by quadruplets)
3) make HighPtTriplets and PixelPair seed-extensions...

MTV
1000 single muon (blue pre5, red WP99p5, orange Wp99p9, black this PR) 
http://innocent.home.cern.ch/innocent/RelVal/Muon10csf99/plots_ootb/effandfake1.pdf
250 ttbar PU200  (blue pre5, red WP99p9, black Wp99p5, orange this PR) 
http://innocent.home.cern.ch/innocent/RelVal/ttbar200PUcsf/plots_ootb/effandfake1.pdf

timing ttbar PU200 tracking only +DQM
900_pre5
```
TimeReport ---------- Event  Summary ---[sec]----
TimeReport       event loop CPU/event = 202.108402
TimeReport      event loop Real/event = 35.805349
TimeReport     sum Streams Real/event = 203.915857
```
this PR
```
TimeReport ---------- Event  Summary ---[sec]----
TimeReport       event loop CPU/event = 146.413452
TimeReport      event loop Real/event = 25.872114
TimeReport     sum Streams Real/event = 148.02895
```




obsoletes #17810