@cmsbuild 2017-03-11
A new Pull Request was created by  @VinInn (Vincenzo Innocente) for master.

It involves the following packages:

RecoPixelVertexing/PixelLowPtUtilities
RecoTracker/FinalTrackSelectors
RecoTracker/IterativeTracking


@perrotta, @cmsbuild, @slava77, @davidlange6 can you please review it and eventually sign? Thanks.
@ghellwig, @makortel, @felicepantaleo, @GiacomoSguazzoni, @rovere, @VinInn, @mschrode, @gpetruc, @dgulhan this is something you requested to watch as well.
@muzaffar, @davidlange6, @smuzaffar you are the release manager for this.

cms-bot commands are listed here cms-sw/cmssw#13028