@ebrondol 2017-03-13
Thanks @VinInn . Can you please describe the legend in the files you have posted?