@slava77 2017-03-16
@VinInn 
what is the purpose of the new files?
They are not used in this PR.

Is it "for upcoming developments"?