@VinInn 2017-03-16
I added the files to allow people to easily test different options changing just the config...