@slava77 2017-03-16
Some plots from PU200 (baseline 900pre6):

some large reduction of "bad" PVs
![all_sign880vsorig_ttbar14tev2023d4timingpu200localwf23034p0c_log10recovertexs_offlineprimaryvertices__reco_obj_zerror](https://cloud.githubusercontent.com/assets/4676718/24011077/f667f28a-0a36-11e7-9704-29b20d9aa7b2.png)

Pretty large reduction in electron seeds (/2.5) with better matching ones retained
![all_sign880vsorig_ttbar14tev2023d4timingpu200localwf23034p0c_recoelectronseeds_electronmergedseeds__reco_obj_drz1](https://cloud.githubusercontent.com/assets/4676718/24011313/b2cd8e30-0a37-11e7-9de8-8cc1d3e9d1cd.png)

downstream we have two times less electronGsfTracks
![all_sign880vsorig_ttbar14tev2023d4timingpu200localwf23034p0c_recogsftracks_electrongsftracks__reco_obj_eta](https://cloud.githubusercontent.com/assets/4676718/24011581/6531714a-0a38-11e7-9cc8-4bb5ec3fd28b.png)


ecal driven electron reduction is mainly in the barrel (apparently mostly fakes; based on MTV GSF eff plots)
![all_sign880vsorig_ttbar14tev2023d4timingpu200localwf23034p0c_recogsfelectrons_ecaldrivengsfelectrons__reco_obj_eta](https://cloud.githubusercontent.com/assets/4676718/24011451/0e8aedb2-0a38-11e7-8b93-c5e9b16ef41a.png)

Inclusive muon distribution is still full of fakes [a part of it is a feature of what's reconstructed/saved, not necessarily fake tracks]
![all_sign880vsorig_ttbar14tev2023d4timingpu200localwf23034p0c_recomuons_muons__reco_obj_eta](https://cloud.githubusercontent.com/assets/4676718/24011649/9d746166-0a38-11e7-9a04-1f449874c3e9.png)

number of generalTracks is down (almost all of this reduction is from tracks with less than 7 hits)
![all_sign880vsorig_ttbar14tev2023d4timingpu200localwf23034p0c_recotracks_generaltracks__reco_obj_eta](https://cloud.githubusercontent.com/assets/4676718/24011911/5fff3738-0a39-11e7-8c30-f833623c5758.png)

PF CH (~high purity tracks) are less affected suggesting relatively high purity here already
![all_sign880vsorig_ttbar14tev2023d4timingpu200localwf23034p0c_recopfcandidates_particleflow__reco_obj_eta43](https://cloud.githubusercontent.com/assets/4676718/24011788/00a2aec8-0a39-11e7-970f-8c0ed7a1557d.png)

.... ..... ..... ......

DQM plots:

wf 21211 (10 muons extended eta range)
muon efficiencies have some visible loss in eta>3.5, while below that the efficiency looks about the same (a bit up/down in a few bins)
<img width="849" alt="wf21211_hp_eff_eta" src="https://cloud.githubusercontent.com/assets/4676718/24014344/d16e1d0a-0a41-11e7-8ec7-c6029d083ce9.png">

In PU200 ttbar
track fake rate is down by x2 or even more in some regions [in line with the supplied MTV plots]
mainly at low pt
<img width="837" alt="wf23034pu200_hp_fake_eta" src="https://cloud.githubusercontent.com/assets/4676718/24014417/1c6ae432-0a42-11e7-97a5-61e5c62f9376.png">
<img width="837" alt="wf23034pu200_hp_fake_eta" src="https://cloud.githubusercontent.com/assets/4676718/24014423/1fef601a-0a42-11e7-9a0e-1b8bdb2267ab.png">

PV fake rate is down
<img width="824" alt="wf23034pu200_offpv_fake_vs_numv" src="https://cloud.githubusercontent.com/assets/4676718/24014438/309c8a46-0a42-11e7-8882-9194162f419b.png">

b-tagging is improved moderately
<img width="839" alt="wf23034pu200_csvv2_l_v_b" src="https://cloud.githubusercontent.com/assets/4676718/24014464/4600f494-0a42-11e7-85a9-7ef9bc7ba5df.png">

.... ..... ..... ......

On the technical side, based on PU200:
- miniAOD output size is down by 5%; RECO (which is actually FEVT) is down by 10%
- total CPU time in reco+miniAOD is reduced by 1.5 from 600 to 400 s/event [in my 10HS06 seconds]
    - iterative tracking "proper" (\*Step\* module names from tracking) is down by a bit more than twice: from 270 s to 120 s/evt 
    -  note that this check is before 4D PV timing went up by x3; after this PR in 91X it's probably heavier than all of tracking
- memory is down by about 200 MB/core (the job includes validation sequences; proper reco may reduce by less)

Overall, looks like a very welcome change.