@kpedro88 2017-03-16
Indeed, a nice improvement both for physics and resource usage