@slava77 2017-03-16
+1

for https://github.com/cms-sw/cmssw/pull/17891	bb2e881
- implemented as described; main improvement are from the cluster shape cuts with some recovery from the pixel doublets
- jenkins tests pass and comparisons with baseline show changes only in 2023 workflows; changes start from tracking as expected
- local tests with 10-mu and TTbarPU200 samples show improvements in behavior, as posted above