@cmsbuild 2017-03-16
This pull request is fully signed and it will be integrated in one of the next master IBs (tests are also fine). This pull request requires discussion in the ORP meeting before it's merged. @muzaffar, @davidlange6, @smuzaffar