@ebrondol 2017-03-17
@slava77 did you also performed the check on D11? Or just a rough idea if my report in PR #17919 is compatible with your report?