@slava77 2017-03-17
On 3/17/17 2:58 AM, ebrondol wrote:
> @slava77 <https://github.com/slava77> did you also performed the check
> on D11? Or just a rough idea if my report in PR #17919
> <https://github.com/cms-sw/cmssw/pull/17919> is compatible with your report?


No, I did not check D11.

Remind me please the significance of it.

From README:
D4 = T3+C2+M1+I1
D11 = T5+C2+M1+I1

and
T3: Phase2 tilted tracker (v3.6.5) w/ phase 2 pixel (v4.0.2.6)
T5: Phase2 tilted tracker (v3.6.5) w/ phase 2 pixel (v4.0.2.5)
{the pixel (v4.0.2.5) has one more FPIX disk}

Is this 8+4 pixel going to become the new default or is it a side-test?



> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/cms-sw/cmssw/pull/17891#issuecomment-287313230>, or
> mute the thread
> <https://github.com/notifications/unsubscribe-auth/AEdcbp-ga4OwKFqSaK53XikbFt946Jkeks5rmllCgaJpZM4MaNHE>.
>