@ebrondol 2017-03-17
As far as I understood, the nomenclature T3 and T5 will remain the same with the idea of having a "validated tracker" (T3) and a "testing tracker" (T5). If T5 will succeed the tests and the validation we are going to do with the next RelVal production, it will become the new default (T3). 
@boudoul @kpedro88 please confirm if I am correct.